Correção delay_ms

+++ ext_libs.c
@@ -71,10 +71,10 @@ double temperatura(uint16_t ADC_RAW_value, int16_t VREF) {
  */
 void delay_ms(uint16_t ms){
  
-  uint8_t i=0;
+  uint16_t i=0;
   uint16_t j=0;
  
-  for(i=0;i<ms*0.1;i++) 
+  for(i=0;i<(double)ms*0.1;i++) 
    for(j=0;j<8335;j++); 
   
 }
